+++ client/src/views/Details/DetailHotel/DetailHotel.jsx
@@ -3,39 +3,10 @@ import { useDispatch, useSelector } from "react-redux";
 import { useParams } from "react-router-dom";
 import { getDetailHotel } from "../../../redux/actions";
 import {loadStripe} from '@stripe/stripe-js';
-import { Elements, CardElement, useStripe, useElements } from "@stripe/react-stripe-js";
-import style from './CheckoutForm.module.css'
+import { Elements } from "@stripe/react-stripe-js";
+import CheckoutForm from "../../../components/CheckoutForm/CheckoutForm"
 
 const stripePromise = loadStripe(import.meta.env.VITE_KEY_STRIPE)
-
-const CheckoutForm = () =>{
-  const stripe = useStripe();
-  const elements = useElements();
-
-  const handleSubmit = async (e) =>{
-    e.preventDefault();
-
-    const {error, paymentMethod} = await stripe.createPaymentMethod({
-      type: 'card',
-      card: elements.getElement(CardElement)
-    });
-
-    if(!error){
-      console.log(paymentMethod);
-    }else{
-      console.log("Error: -> ",error);
-    }
-  }
-  
-  return (
-    <form onSubmit={handleSubmit} className={style.paymentForm}>
-      <CardElement />
-      <button>
-        Reservar
-      </button>
-    </form>
-  )
-}
 
 const DetailHotel = () => {
   const dispatch = useDispatch();
